rename stage app stacks
