+++ typescript/aws-codepipeline-ecs-lambda/lib/pipeline-stack.ts
@@ -1,7 +1,7 @@
 import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { CodePipeline, CodePipelineSource, ManualApprovalStep, ShellStep, Wave } from 'aws-cdk-lib/pipelines';
-import { pipelineAppStage } from './app-stage';
+import { pipelineAppStage } from './stage-app';
 
 export class pipelineStack extends cdk.Stack {
   constructor(scope: Construct, id: string, props?: cdk.StackProps) {
