+++ typescript/aws-codepipeline-ecs-lambda/lib/stage-app-async-lambda-stack.ts
