+++ typescript/aws-codepipeline-ecs-lambda/lib/stage-app-ecs-fargate-stack.ts
