+++ typescript/aws-codepipeline-ecs-lambda/lib/stage-app-lambda-api-stack.ts
